+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/LightDefinition.cs
@@ -82,6 +82,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
         public float   diffuseDimmer;
         public float   specularDimmer;
+        public int     interactsWithSky;        // bool...
     };
 
     [GenerateHLSL(PackingRules.Exact, false)]